Really moved components to a components dir
